backend/requirements.txt
beautifulsoup4
requests
gym
numpy
modal
arcgis
Flask
python-dotenv
